requirements.txt
Django>=5.0,<6.0
requests>=2.0
Pillow>=9.0
